+++ src/Providers/Jspm.cs
@@ -28,6 +28,11 @@ namespace PackageInstaller
             get { return _icon; }
         }
 
+        public override string DefaultArguments
+        {
+            get { return VSPackage.Settings.JspmArguments; }
+        }
+
         public override async Task<IEnumerable<string>> GetPackages(string term = null)
         {
             string file = Path.Combine(Path.GetTempPath(), "jspm-registry.txt");
@@ -41,12 +46,11 @@ namespace PackageInstaller
             return await Task.FromResult(Enumerable.Empty<string>());
         }
 
-        public override async Task<bool> InstallPackage(Project project, string packageName, string version)
+        public override async Task<bool> InstallPackage(Project project, string packageName, string version, string args = null)
         {
-            if (!string.IsNullOrEmpty(version))
-                packageName += $"@{version}";
+            string installArgs = GetInstallArguments(packageName, version);
 
-            string arg = $"/c jspm install {packageName} {VSPackage.Settings.JspmArugments}";
+            string arg = $"/c {installArgs} {args}";
             string cwd = project.GetRootFolder();
             string json = Path.Combine(cwd, "package.json");
 
@@ -65,6 +69,16 @@ namespace PackageInstaller
             {
                 return await ShowConsole(arg, cwd);
             }
+        }
+
+        public override string GetInstallArguments(string name, string version)
+        {
+            string args = $"jspm install {name}";
+
+            if (!string.IsNullOrEmpty(version))
+                args = $"{args}@{version}";
+
+            return args;
         }
 
         private bool IsJspmConfigured(string packageJsonFile)
@@ -167,7 +181,7 @@ namespace PackageInstaller
             var doc = JObject.Parse(json);
 
             return doc.Children<JProperty>()
-                      .OrderBy(prop => prop.Name)
+                      .OrderBy(prop => prop.Name, new PackageNameComparer())
                       .Select(prop => prop.Name);
         }
     }