+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Roadmap
 
-- [ ] Bower: Convert to new Azure based registry
+- [x] Add textbox for arguments
+- [x] Better sorting of package names
 - [ ] Implement better auto-completion mechanism
 
 Features that have a checkmark are complete and available for